feat(api): 添加系统管理相关API接口
- 新增算法管理API，包括增删改查功能
- 新增数据集管理API，支持树形列表和表单操作
- 新增部门管理API，包含部门列表和下拉选项
- 新增字典管理API，支持字典项和字典类型分页查询
- 新增文件管理API，支持文件上传、下载和删除
- 新增数据项文件管理API，支持图片上传功能
- 扩展API组结构体，集成多个系统服务接口
- 初始化各服务模块的依赖注入配置

+++ dehaze-go/service/sys_user.go
@@ -46,7 +46,7 @@ func (userService *UserService) GetUserAuthInfo(username string) (userAuthInfo *
 	userAuthInfo.Nickname = user.Nickname
 	userAuthInfo.Password = user.Password
 	userAuthInfo.Status = user.Status
-	userAuthInfo.DeptId = user.DeptID
+	userAuthInfo.DeptId = user.DeptID  // 修复类型转换
 
 	// 如果查询到用户信息
 	if userAuthInfo.UserId != 0 {
@@ -96,7 +96,7 @@ func (userService *UserService) GetUserAuthInfo(username string) (userAuthInfo *
 			}
 
 			if dataScope != nil {
-				userAuthInfo.DataScope = int8(*dataScope)
+				userAuthInfo.DataScope = *dataScope
 			}
 		}
 	}
